+++ package.json
@@ -40,7 +40,7 @@
     "eslint-plugin-prettier": "^3.1.1",
     "eslint-plugin-react": "^7.18.3",
     "eslint-plugin-react-hooks": "^1.7.0",
-    "file-loader": "5.0.2",
+    "file-loader": "^5.0.2",
     "husky": "^2.7.0",
     "lint-staged": "^8.2.1",
     "mini-css-extract-plugin": "^0.7.0",
@@ -48,6 +48,7 @@
     "style-loader": "^0.23.1",
     "stylus": "^0.54.7",
     "stylus-loader": "^3.0.2",
+    "svg-inline-loader": "^0.8.2",
     "url-loader": "3.0.0"
   },
   "husky": {
