+++ src/components/Nav.jsx
@@ -3,7 +3,7 @@ import '../styles/components/Nav.styl';
 
 function Nav() {
   return (
-    <nav className="Nav">
+    <nav className="Nav" id="Nav">
       <ul className="Nav__list">
         <li className="Nav_element">
           <a href="/">TECHNOLOGY</a>
