+++ src/styles/components/Header.styl
@@ -1,7 +1,20 @@
 .Header
   width 100%
   display grid
-  grid-template-columns: 40px 1fr 30px
-
+  grid-template-columns: 100px 50px
+  justify-content: space-between
+  align-content: center
+  position fixed
+  background white
+  padding 10px 0
+  @media screen and (min-width 1024px)
+    grid-template-columns: 100px 1fr
+    justify-content: space-between
+  &__logoFigure
+    display grid
+    justify-content: center
+    align-content: center
 .Hamburger
-  display block
+  display grid
+  align-content: center
+  z-index 20
